svix-cli(wizard): split into different files
I was checking some ratatui projects from https://ratatui.rs/showcase/apps/ to see how they're usually organized and I noticed there's usually a file for the state machine, other for things that draw in the TUI and other for the reusable components
