+++ svix-cli/src/cmds/wizard/tui/mod.rs
@@ -3,27 +3,30 @@
 //! The steps are a checklist you work down: each one builds on the one before it, so the
 //! wizard only moves forward, ticking steps off as they're done. The only key that
 //! changes anything on your account is Enter.
+//!
+//! This file owns the state and what the keys do to it; [`render`] draws it, and
+//! [`widgets`] holds the drawing helpers the steps share.
+
+mod render;
+mod widgets;
 
 use std::{
     cell::Cell,
+    sync::LazyLock,
     time::{Duration, Instant},
 };
 
 use ratatui::{
     crossterm::event::{self, Event, KeyCode, KeyEvent, KeyEventKind, KeyModifiers},
-    layout::{Constraint, Layout, Rect},
-    style::{Color, Modifier, Style},
-    text::{Line, Span},
-    widgets::{Paragraph, Wrap},
-    DefaultTerminal, Frame,
+    DefaultTerminal,
 };
 use svix::api::Svix;
 
+use self::widgets::copy_to_clipboard;
 use super::{
-    create_application,
-    highlight::{background, highlight, Syntax},
-    install_skill, portal_url, sample_message, send_message, Language, Quickstart, QuickstartMode,
-    SampleMessage, SkillScope, DEFAULT_SERVER_URL, LANGUAGES, SKILL_NAMES,
+    create_application, highlight::Syntax, install_skill, portal_url, sample_message, send_message,
+    Language, Quickstart, QuickstartMode, SampleMessage, SkillScope, DEFAULT_SERVER_URL, LANGUAGES,
+    SKILL_NAMES,
 };
 use crate::{
     cmds::login::{self, DashboardLogin},
@@ -129,6 +132,12 @@ enum Auth {
     Done,
 }
 
+impl Default for Auth {
+    fn default() -> Self {
+        Self::Choosing { selected: 0 }
+    }
+}
+
 const AUTH_CHOICES: &[&str] = &["Log in through dashboard.svix.com", "Paste an auth token"];
 const MODE_CHOICES: &[&str] = &[
     "Continue manually (I'll walk through the steps myself)",
@@ -139,6 +148,11 @@ const SCOPE_CHOICES: &[&str] = &[
     "Globally, for every project (~/.agents/skills)",
 ];
 
+/// The names in `LANGUAGES`, in a shape the choice lists can borrow.
+static LANGUAGE_NAMES: LazyLock<Vec<&'static str>> =
+    LazyLock::new(|| LANGUAGES.iter().map(|language| language.name).collect());
+
+#[derive(Default)]
 struct App {
     cfg: Config,
     client: Option<Svix>,
@@ -186,30 +200,13 @@ impl App {
                 .then(|| crate::get_client(&cfg).ok())
                 .flatten(),
             cfg,
-            step: 0,
-            language: 0,
-            scroll: 0,
-            quit: false,
             width: Cell::new(80),
             auth: if authenticated {
                 Auth::Done
             } else {
-                Auth::Choosing { selected: 0 }
+                Auth::default()
             },
-            mode: None,
-            mode_selected: 0,
-            scope: None,
-            scope_selected: 0,
-            skills_installed: false,
-            skills_done: 0,
-            message: None,
-            quickstart: None,
-            sent: None,
-            portal: None,
-            pending: None,
-            busy: None,
-            status: None,
-            error: None,
+            ..Self::default()
         }
     }
 
@@ -440,20 +437,15 @@ impl App {
     }
 
     /// The options the current step is asking the user to pick from, if it is.
-    fn choices(&self) -> Option<(Vec<&'static str>, usize)> {
+    fn choices(&self) -> Option<(&'static [&'static str], usize)> {
         match (self.step(), &self.auth) {
-            (Step::Auth, Auth::Choosing { selected }) => Some((AUTH_CHOICES.to_vec(), *selected)),
-            (Step::Mode, _) if self.mode.is_none() => {
-                Some((MODE_CHOICES.to_vec(), self.mode_selected))
-            }
+            (Step::Auth, Auth::Choosing { selected }) => Some((AUTH_CHOICES, *selected)),
+            (Step::Mode, _) if self.mode.is_none() => Some((MODE_CHOICES, self.mode_selected)),
             // The agent path stays on this step to ask where the skills should go.
             (Step::Mode, _) if self.mode == Some(QuickstartMode::Agent) && self.scope.is_none() => {
-                Some((SCOPE_CHOICES.to_vec(), self.scope_selected))
+                Some((SCOPE_CHOICES, self.scope_selected))
             }
-            (Step::Language, _) => Some((
-                LANGUAGES.iter().map(|language| language.name).collect(),
-                self.language,
-            )),
+            (Step::Language, _) => Some((LANGUAGE_NAMES.as_slice(), self.language)),
             _ => None,
         }
     }
@@ -468,7 +460,8 @@ impl App {
             Step::Auth => self.auth = Auth::Choosing { selected: next },
             Step::Language => self.language = next,
             Step::Mode if self.mode.is_some() => self.scope_selected = next,
-            _ => self.mode_selected = next,
+            Step::Mode => self.mode_selected = next,
+            _ => {}
         }
     }
 
@@ -605,171 +598,6 @@ impl App {
         });
     }
 
-    /// Reprints the ids and URLs once the alternate screen is gone.
-    fn print_summary(&self) {
-        let Some(qs) = &self.quickstart else {
-            return;
-        };
-
-        println!("Quickstart summary:");
-        println!("  Application:  {} ({})", qs.app.name, qs.app.id);
-        if let Some(uid) = &qs.app.uid {
-            println!("  Uid:          {uid}");
-        }
-        println!("  Endpoint:     {} -> {}", qs.endpoint.id, qs.endpoint.url);
-        println!("  Play inbox:   {}", qs.play_view_url);
-        if let Some(id) = &self.sent {
-            println!("  Message sent: {id}");
-        }
-        if let Some(url) = &self.portal {
-            println!("  App portal:   {url}");
-        }
-
-        println!("\nFrom here you can:");
-        println!("  - Add more event types:  `{BIN_NAME} event-type create {{...}}`");
-        println!(
-            "  - Forward webhooks to a local server:  `{BIN_NAME} listen http://localhost:8000/webhook`"
-        );
-        println!("  - Read the docs:  https://docs.svix.com\n");
-    }
-}
-
-// --- rendering -----------------------------------------------------------------------
-
-// Prose and code keep the terminal's own foreground colour, and anything secondary is
-// dimmed rather than greyed: a fixed grey is unreadable on a light background.
-const HEADING: Style = Style::new().fg(Color::Cyan).add_modifier(Modifier::BOLD);
-const VALUE: Style = Style::new().fg(Color::Green);
-const DIM: Style = Style::new().add_modifier(Modifier::DIM);
-const TEXT: Style = Style::new();
-
-impl App {
-    fn render(&self, frame: &mut Frame) {
-        // The step list grows downwards as steps are reached, and the body gets the rest.
-        // Nothing sits beside the body, so a wrapped URL can be selected without picking
-        // up anything else on the row.
-        let steps_height = (self.step + 3) as u16;
-        let [title, steps, body, footer] = Layout::vertical([
-            Constraint::Length(1),
-            Constraint::Length(steps_height),
-            Constraint::Min(0),
-            Constraint::Length(1),
-        ])
-        .areas(frame.area());
-
-        self.render_title(frame, title);
-        self.render_steps(frame, steps);
-        self.render_body(frame, body);
-        self.render_footer(frame, footer);
-    }
-
-    fn render_title(&self, frame: &mut Frame, area: Rect) {
-        frame.render_widget(Line::styled("Svix quickstart", HEADING), area);
-    }
-
-    /// The steps reached so far: done ones keep a check, the current one is highlighted.
-    /// The ones still ahead are left out until they're reached.
-    fn render_steps(&self, frame: &mut Frame, area: Rect) {
-        let mut lines: Vec<Line> = STEPS
-            .iter()
-            .enumerate()
-            .take(self.step + 1)
-            .map(|(index, step)| {
-                let (marker, style) = if self.done(index) {
-                    ("✓", VALUE)
-                } else if index == self.step {
-                    ("▸", HEADING)
-                } else {
-                    ("·", DIM)
-                };
-
-                Line::styled(format!("{marker} {}", step.title()), style)
-            })
-            .collect();
-
-        // A rule under the list, standing in for the panel border the body no longer has.
-        lines.push(Line::styled("─".repeat(area.width as usize), DIM));
-        lines.push(Line::from(""));
-
-        frame.render_widget(Paragraph::new(lines), area);
-    }
-
-    fn render_body(&self, frame: &mut Frame, area: Rect) {
-        // The sample boxes are drawn as text, so they have to be built for this width.
-        let width = area.width;
-        self.width.set(width);
-        let mut lines = self.step_lines(width);
-
-        if let Some(busy) = self.busy {
-            lines.push(Line::from(""));
-            // Bold rather than yellow, which washes out on a light background.
-            lines.push(Line::styled(
-                busy,
-                Style::new().add_modifier(Modifier::BOLD),
-            ));
-        }
-        if let Some(status) = &self.status {
-            lines.push(Line::from(""));
-            lines.push(Line::styled(status.clone(), VALUE));
-        }
-        if let Some(error) = &self.error {
-            lines.push(Line::from(""));
-            lines.push(Line::styled(
-                format!("Error: {error}"),
-                Style::new().fg(Color::Red),
-            ));
-            lines.push(Line::styled("Press r to try again.", DIM));
-        }
-
-        // No border and no padding, so selecting a wrapped URL picks up the URL alone.
-        let paragraph = Paragraph::new(lines)
-            .wrap(Wrap { trim: false })
-            .scroll((self.scroll, 0));
-
-        frame.render_widget(paragraph, area);
-    }
-
-    fn render_footer(&self, frame: &mut Frame, area: Rect) {
-        // Typing a token swallows the usual keys, so it gets its own footer.
-        if matches!((self.step(), &self.auth), (Step::Auth, Auth::Typing { .. })) {
-            let keys = "type your token  ·  enter save  ·  esc back";
-            frame.render_widget(Line::styled(keys, DIM), area);
-            return;
-        }
-
-        if matches!(
-            (self.step(), &self.auth),
-            (Step::Auth, Auth::Waiting { .. })
-        ) {
-            let keys = "waiting for the browser  ·  q quit";
-            frame.render_widget(Line::styled(keys, DIM), area);
-            return;
-        }
-
-        let mut keys = match self.step() {
-            Step::Send if self.sent.is_none() => vec!["enter send".to_owned()],
-            Step::Done => vec!["enter finish".to_owned()],
-            _ if self.choices().is_some() => vec!["enter select".to_owned()],
-            _ => vec!["enter continue".to_owned()],
-        };
-
-        if self.choices().is_some() {
-            keys.push("↑↓/j/k choose".to_owned());
-        }
-        if self.open_url().is_some() {
-            keys.push("o open".to_owned());
-        }
-        if let Some((label, _)) = self.copy_target() {
-            keys.push(format!("c copy {label}"));
-        }
-        if self.choices().is_none() {
-            keys.push("↑↓/j/k scroll".to_owned());
-        }
-        keys.push("q quit".to_owned());
-
-        frame.render_widget(Line::styled(keys.join("  ·  "), DIM), area);
-    }
-
     /// The code sample the current step shows, as its title and text.
     fn sample_text(&self) -> Option<(&'static str, Syntax, String)> {
         let server_url = self.cfg.server_url().unwrap_or(DEFAULT_SERVER_URL);
@@ -824,445 +652,33 @@ impl App {
         });
     }
 
-    fn step_lines(&self, width: u16) -> Vec<Line<'static>> {
-        match self.step() {
-            Step::Auth => self.auth_lines(),
-            Step::Mode => self.mode_lines(),
-            Step::Language => self.language_lines(),
-            Step::Application => self.application_lines(width),
-            Step::Send => self.send_lines(width),
-            Step::Portal => self.portal_lines(width),
-            Step::Done => self.done_lines(),
-        }
-    }
-
-    fn auth_lines(&self) -> Vec<Line<'static>> {
-        let mut lines =
-            wrap_text("The quickstart works against your Svix account, so it needs an API token.");
-        lines.push(Line::from(""));
-
-        match &self.auth {
-            Auth::Choosing { selected } => lines.extend(choices(AUTH_CHOICES, *selected)),
-            Auth::Typing { token } => {
-                lines.extend(wrap_text("Paste your auth token and press enter:"));
-                lines.push(Line::from(""));
-                lines.push(Line::from(vec![
-                    Span::styled("> ", DIM),
-                    Span::styled(token.clone(), VALUE),
-                    Span::styled("_", DIM),
-                ]));
-                lines.push(Line::from(""));
-                lines.push(Line::styled("esc to go back to the other options", DIM));
-            }
-            Auth::Waiting { session, .. } => {
-                lines.extend(wrap_text(
-                    "Approve the login in your browser, then come back here.",
-                ));
-                lines.push(Line::from(""));
-                lines.push(field("Verification code", &session.code));
-                lines.push(Line::from(""));
-                lines.extend(wrap_text("If the browser didn't open, use this URL:"));
-                lines.push(Line::styled(session.url.clone(), VALUE));
-                lines.push(Line::from(""));
-                lines.push(Line::styled("Waiting for approval...", DIM));
-            }
-            Auth::Done => {
-                lines.push(Line::styled("You're authenticated with Svix.", VALUE));
-                if let Some(url) = self.cfg.server_url() {
-                    lines.push(Line::from(""));
-                    lines.push(field("Server", url));
-                }
-                lines.push(Line::from(""));
-                lines.push(Line::styled("Press enter to continue.", HEADING));
-            }
-        }
-
-        lines
-    }
-
-    fn mode_lines(&self) -> Vec<Line<'static>> {
-        let mut lines = wrap_text(
-            "You can walk through the rest of the quickstart yourself, or hand it to a coding \
-             agent: picking the agent installs the Svix skills and hands over to them.",
-        );
-        lines.push(Line::from(""));
-
-        match self.mode {
-            Some(QuickstartMode::Manual) => {
-                lines.push(Line::styled("Continuing manually.", VALUE));
-            }
-            Some(QuickstartMode::Agent) => {
-                lines.push(Line::styled("An agent will take it from here.", VALUE));
-                lines.push(Line::from(""));
-                lines.extend(self.skills_lines());
-            }
-            None => lines.extend(choices(MODE_CHOICES, self.mode_selected)),
-        }
-
-        lines
-    }
-
-    /// The tail of the agent path: the skills being installed, where they'll go, and how
-    /// far through the run is.
-    fn skills_lines(&self) -> Vec<Line<'static>> {
-        let mut lines =
-            wrap_text("The skills are installed for every coding agent found on this machine:");
-        lines.push(Line::from(""));
-        lines.extend(self.install_lines());
-        lines.push(Line::from(""));
-
-        // Once the scope is picked there's nothing left to ask: either a run is in flight,
-        // or one of them failed and the error below says so. A finished run doesn't get
-        // here at all, since the wizard quits to hand over to the agent.
-        if self.scope.is_some() {
-            return lines;
-        }
-
-        lines.extend(wrap_text("Where should they go?"));
-        lines.push(Line::from(""));
-        lines.extend(choices(SCOPE_CHOICES, self.scope_selected));
-
-        lines
-    }
-
-    /// One line per skill, marked with how far the run has got.
-    ///
-    /// Nothing is running until the scope is picked, so until then every skill is listed
-    /// as queued rather than the first one looking like it's already going.
-    fn install_lines(&self) -> Vec<Line<'static>> {
-        SKILL_NAMES
-            .iter()
-            .enumerate()
-            .map(|(index, skill)| {
-                let running =
-                    self.scope.is_some() && self.busy.is_some() && index == self.skills_done;
-                let (marker, style) = if index < self.skills_done {
-                    ("✓", VALUE)
-                } else if running {
-                    ("▸", HEADING)
-                } else {
-                    ("·", DIM)
-                };
-
-                Line::styled(format!("{marker} Installing {skill} using npx"), style)
-            })
-            .collect()
-    }
-
-    fn language_lines(&self) -> Vec<Line<'static>> {
-        let mut lines = wrap_text(
-            "Pick the language you'll be integrating in. The code samples in the rest of the \
-             quickstart are shown in it.",
-        );
-        lines.push(Line::from(""));
-        lines.extend(choices(
-            &LANGUAGES
-                .iter()
-                .map(|language| language.name)
-                .collect::<Vec<_>>(),
-            self.language,
-        ));
-
-        lines
-    }
-
-    fn application_lines(&self, width: u16) -> Vec<Line<'static>> {
-        let mut lines = wrap_text(
-            "A consumer application defines where your messages are sent. Usually you'll want \
-             one application for each of your customers.",
-        );
-
+    /// Reprints the ids and URLs once the alternate screen is gone.
+    fn print_summary(&self) {
         let Some(qs) = &self.quickstart else {
-            return lines;
+            return;
         };
 
-        lines.push(Line::from(""));
-        lines.extend([field("Application", &qs.app.name), field("Id", &qs.app.id)]);
+        println!("Quickstart summary:");
+        println!("  Application:  {} ({})", qs.app.name, qs.app.id);
         if let Some(uid) = &qs.app.uid {
-            lines.push(field("Uid", uid));
+            println!("  Uid:          {uid}");
         }
-        lines.push(Line::from(""));
-        lines.extend(wrap_text(
-            "It also got an example endpoint pointing at a Svix Play inbox, so you can preview \
-             the message sent here. Your customers add their own endpoints in the \
-             app portal.",
-        ));
-        lines.push(Line::from(""));
-        lines.extend([
-            field("Endpoint", &qs.endpoint.id),
-            field("Delivers to", &qs.endpoint.url),
-        ]);
-        if let Some(channel) = &qs.channel {
-            lines.push(field("Channel", channel));
+        println!("  Endpoint:     {} -> {}", qs.endpoint.id, qs.endpoint.url);
+        println!("  Play inbox:   {}", qs.play_view_url);
+        if let Some(id) = &self.sent {
+            println!("  Message sent: {id}");
+        }
+        if let Some(url) = &self.portal {
+            println!("  App portal:   {url}");
         }
 
-        lines.extend(self.code_lines(width));
-
-        lines
-    }
-
-    /// How to send a message to the application from your own code, in the language picked
-    /// on the previous step.
-    fn code_lines(&self, width: u16) -> Vec<Line<'static>> {
-        let language = self.language();
-
-        let mut lines = vec![
-            Line::from(""),
-            Line::styled(format!("1. Install the SDK ({})", language.name), HEADING),
-            Line::from(""),
-        ];
-        lines.extend(highlight(Syntax::Shell, language.install));
-        lines.push(Line::from(""));
-        lines.push(Line::styled("2. Export your API token", HEADING));
-        lines.push(Line::from(""));
-        lines.extend(highlight(
-            Syntax::Shell,
-            "export SVIX_AUTH_TOKEN='<your-token>'",
-        ));
-        lines.push(Line::from(""));
-        lines.extend(wrap_text(
-            "Create environment-specific tokens in the dashboard: https://dashboard.svix.com",
-        ));
-        lines.push(Line::from(""));
-        lines.push(Line::styled("3. Send a message from your code", HEADING));
-        lines.push(Line::from(""));
-        lines.extend(wrap_text(
-            "Put this where the event actually happens. The same spot you'd log it or fire an \
-             internal event.",
-        ));
-        lines.push(Line::from(""));
-
-        match self.sample_text() {
-            Some((title, syntax, text)) => {
-                lines.extend(sample(title, syntax, &text, width));
-                lines.push(Line::from(""));
-                lines.extend(wrap_text(
-                    "In your real integration you'd create one application per customer and use \
-                     that customer's application id (or uid) here, instead of the hardcoded one \
-                     above.",
-                ));
-            }
-            _ => lines.push(Line::styled(
-                "Waiting for the application to be created...",
-                DIM,
-            )),
-        }
-
-        lines
-    }
-
-    fn send_lines(&self, width: u16) -> Vec<Line<'static>> {
-        let mut lines = wrap_text(
-            "This sends the exact message from the previous step, so you can see it delivered \
-             before wiring up your own code.",
+        println!("\nFrom here you can:");
+        println!("  - Add more event types:  `{BIN_NAME} event-type create {{...}}`");
+        println!(
+            "  - Forward webhooks to a local server:  `{BIN_NAME} listen http://localhost:8000/webhook`"
         );
-        lines.push(Line::from(""));
-
-        let Some(msg) = &self.message else {
-            return lines;
-        };
-
-        lines.push(field("Event type", &msg.event_type));
-        lines.push(Line::from(""));
-        if let Some((title, syntax, text)) = self.sample_text() {
-            lines.extend(sample(title, syntax, &text, width));
-        }
-        lines.push(Line::from(""));
-
-        match (&self.sent, &self.quickstart) {
-            (Some(id), Some(qs)) => {
-                lines.push(field("Sent", id));
-                lines.push(Line::from(""));
-                lines.extend(wrap_text(
-                    "It was delivered to the example endpoint. Press o to open the Svix Play \
-                     inbox and see the request Svix made:",
-                ));
-                lines.push(Line::styled(qs.play_view_url.clone(), VALUE));
-                lines.push(Line::from(""));
-                lines.extend(wrap_text(&format!(
-                    "Every attempt is recorded: `{BIN_NAME} message-attempt list-by-msg {} {id}`",
-                    qs.app.id
-                )));
-            }
-            (None, Some(_)) => {
-                lines.push(Line::styled("Press enter to send it.", HEADING));
-                lines.push(Line::styled(UNLOCK_HINT, DIM));
-            }
-            _ => {}
-        }
-
-        lines
+        println!("  - Read the docs:  https://docs.svix.com\n");
     }
-
-    fn portal_lines(&self, width: u16) -> Vec<Line<'static>> {
-        let mut lines = wrap_text(
-            "In production, your webhook consumers add their own endpoints inside your product \
-             using the pre-built, embeddable app portal. The endpoint and message from the \
-             previous steps are already in there.",
-        );
-        lines.push(Line::from(""));
-
-        match &self.portal {
-            Some(url) => {
-                lines.extend(wrap_text("Press o to open this one-time link:"));
-                lines.push(Line::styled(url.clone(), VALUE));
-            }
-            None => lines.push(Line::styled("No link yet. Press r to generate one.", DIM)),
-        }
-
-        lines.push(Line::from(""));
-        lines.extend(wrap_text(
-            "Links like that are short-lived, so you mint them on demand from your backend and \
-             link to them from your own dashboard:",
-        ));
-        lines.push(Line::from(""));
-
-        if let Some((title, syntax, text)) = self.sample_text() {
-            lines.extend(sample(title, syntax, &text, width));
-        }
-        if let Some(qs) = &self.quickstart {
-            lines.push(Line::from(""));
-            lines.extend(wrap_text(&format!(
-                "Or from the CLI: `{BIN_NAME} authentication app-portal-access {}`",
-                qs.app.id
-            )));
-        }
-
-        lines
-    }
-
-    fn done_lines(&self) -> Vec<Line<'static>> {
-        let mut lines = wrap_text("That's the quickstart. From here you can:");
-        lines.push(Line::from(""));
-        lines.extend([
-            Line::styled(
-                format!("  Add more event types:  {BIN_NAME} event-type create {{...}}"),
-                VALUE,
-            ),
-            Line::styled(
-                format!(
-                    "  Forward webhooks locally:  {BIN_NAME} listen http://localhost:8000/webhook"
-                ),
-                VALUE,
-            ),
-            Line::styled("  Read the docs:  https://docs.svix.com", VALUE),
-        ]);
-        lines.push(Line::from(""));
-        lines.extend(wrap_text(
-            "Press q to leave. The ids and links from these steps get printed on the way out.",
-        ));
-
-        lines
-    }
-}
-
-/// A `label: value` line, with the value in the colour used for things worth copying.
-fn field(label: &str, value: &str) -> Line<'static> {
-    Line::from(vec![
-        Span::styled(format!("{label}: "), DIM),
-        Span::styled(value.to_owned(), VALUE),
-    ])
-}
-
-/// A highlighted code sample, drawn as text so it scrolls with everything else.
-///
-/// The block is a rule above and below plus the highlighting theme's background, with no
-/// side borders: selecting the sample with the mouse then picks up the code alone.
-fn sample(title: &str, syntax: Syntax, text: &str, width: u16) -> Vec<Line<'static>> {
-    let width = (width as usize).max(8);
-    let rule = background().add_modifier(Modifier::DIM);
-
-    let heading = format!("─ {title} ");
-    let mut lines = vec![Line::styled(
-        format!(
-            "{heading}{}",
-            "─".repeat(width.saturating_sub(heading.chars().count()))
-        ),
-        rule,
-    )];
-
-    lines.extend(
-        highlight(syntax, text)
-            .into_iter()
-            .map(|line| fit(line, width).style(background())),
-    );
-
-    lines.push(Line::styled("─".repeat(width), rule));
-
-    lines
-}
-
-/// Truncates or pads a line to exactly `width` cells.
-fn fit(line: Line<'static>, width: usize) -> Line<'static> {
-    let mut spans = Vec::new();
-    let mut used = 0;
-
-    for span in line.spans {
-        if used >= width {
-            break;
-        }
-        let content = span.content.to_string();
-        let len = content.chars().count();
-
-        if used + len <= width {
-            used += len;
-            spans.push(Span::styled(content, span.style));
-        } else {
-            let kept: String = content.chars().take(width - used - 1).collect();
-            spans.push(Span::styled(format!("{kept}…"), span.style));
-            used = width;
-        }
-    }
-
-    if used < width {
-        spans.push(Span::raw(" ".repeat(width - used)));
-    }
-
-    Line::from(spans)
-}
-
-/// Copies `text` through the terminal itself with OSC 52.
-///
-/// Going through the terminal is what makes this work over SSH and in a multiplexer,
-/// where a clipboard library would reach for the wrong machine's clipboard.
-fn copy_to_clipboard(text: &str) -> std::io::Result<()> {
-    use std::io::Write as _;
-
-    use base64::{engine::general_purpose::STANDARD, Engine as _};
-
-    let sequence = format!("\x1b]52;c;{}\x07", STANDARD.encode(text));
-    // tmux swallows escape sequences it isn't told to pass through.
-    let sequence = match std::env::var_os("TMUX") {
-        Some(_) => format!("\x1bPtmux;\x1b{sequence}\x1b\\"),
-        None => sequence,
-    };
-
-    let mut stdout = std::io::stdout();
-    stdout.write_all(sequence.as_bytes())?;
-    stdout.flush()
-}
-
-/// A pick-one list, rendered inline so it scrolls with the rest of the step.
-fn choices(options: &[&'static str], selected: usize) -> Vec<Line<'static>> {
-    options
-        .iter()
-        .enumerate()
-        .map(|(index, option)| {
-            if index == selected {
-                Line::from(vec![
-                    Span::styled("› ", HEADING),
-                    Span::styled(*option, HEADING),
-                ])
-            } else {
-                Line::styled(format!("  {option}"), DIM)
-            }
-        })
-        .collect()
-}
-
-/// Prose is wrapped by the paragraph widget, so this just owns the text.
-fn wrap_text(text: &str) -> Vec<Line<'static>> {
-    vec![Line::styled(text.to_owned(), TEXT)]
 }
 
 #[cfg(test)]
@@ -1274,6 +690,10 @@ mod tests {
     };
 
     use super::*;
+
+    fn step_index(step: Step) -> usize {
+        STEPS.iter().position(|s| *s == step).expect("a known step")
+    }
 
     /// An app past the login step, which is where most of the wizard's behaviour lives.
     fn app() -> App {
@@ -1418,11 +838,7 @@ mod tests {
         assert_eq!(app.step, 4);
 
         // None of the keys that used to go back do anything now.
-        for key in [
-            KeyCode::Left,
-            KeyCode::Char('h'),
-            KeyCode::BackTab,
-        ] {
+        for key in [KeyCode::Left, KeyCode::Char('h'), KeyCode::BackTab] {
             let step = app.step;
             app.on_key(KeyEvent::from(key));
             assert_eq!(app.step, step, "{key:?} shouldn't move between steps");
@@ -1471,10 +887,7 @@ mod tests {
     fn the_app_portal_stays_shut_until_the_message_is_sent() {
         let mut app = app();
 
-        let send = STEPS
-            .iter()
-            .position(|s| *s == Step::Send)
-            .expect("send step");
+        let send = step_index(Step::Send);
         app.step = send;
 
         // Enter sends rather than moving on, and only sending gets you to the portal.
@@ -1496,10 +909,7 @@ mod tests {
             payload: serde_json::json!({ "id": "invoice_WF7WtC" }),
         });
 
-        app.step = STEPS
-            .iter()
-            .position(|s| *s == Step::Application)
-            .expect("application step");
+        app.step = step_index(Step::Application);
         let (label, text) = app.copy_target().expect("something to copy");
         assert_eq!(label, "sample");
         assert!(
@@ -1510,10 +920,7 @@ mod tests {
         assert!(app.status.is_some(), "copying says what it did");
 
         // Steps with a link copy that instead, matching what `o` opens.
-        app.step = STEPS
-            .iter()
-            .position(|s| *s == Step::Portal)
-            .expect("portal step");
+        app.step = step_index(Step::Portal);
         app.portal = Some("https://app.svix.com/login#key=abc".to_owned());
         let (label, text) = app.copy_target().expect("something to copy");
         assert_eq!(label, "link");
@@ -1523,20 +930,8 @@ mod tests {
     #[test]
     fn steps_ahead_of_the_current_one_are_hidden() {
         let mut app = app();
-        let mut terminal = Terminal::new(TestBackend::new(80, 24)).expect("terminal");
 
-        let rendered = |app: &App, terminal: &mut Terminal<TestBackend>| -> String {
-            terminal.draw(|frame| app.render(frame)).expect("draw");
-            terminal
-                .backend()
-                .buffer()
-                .content()
-                .iter()
-                .map(|cell| cell.symbol())
-                .collect()
-        };
-
-        let first = rendered(&app, &mut terminal);
+        let first = screen(&app);
         assert!(first.contains("Authenticate"), "the current step is listed");
         assert!(
             !first.contains("App portal"),
@@ -1544,7 +939,7 @@ mod tests {
         );
 
         app.step = STEPS.len() - 1;
-        let last = rendered(&app, &mut terminal);
+        let last = screen(&app);
         assert!(last.contains("App portal"), "reaching a step lists it");
     }
 
@@ -1578,10 +973,7 @@ mod tests {
     fn the_language_step_is_a_list_you_pick_from() {
         let mut app = app();
 
-        app.step = STEPS
-            .iter()
-            .position(|s| *s == Step::Language)
-            .expect("language step");
+        app.step = step_index(Step::Language);
         app.on_key(KeyEvent::from(KeyCode::Down));
         assert_eq!(app.language, 1);
         app.on_key(KeyEvent::from(KeyCode::Up));
@@ -1592,10 +984,7 @@ mod tests {
         assert_eq!(app.scroll, 0, "the vertical keys choose here, not scroll");
 
         // Off the language step they scroll instead, and nothing changes the language.
-        app.step = STEPS
-            .iter()
-            .position(|s| *s == Step::Send)
-            .expect("send step");
+        app.step = step_index(Step::Send);
         app.on_key(KeyEvent::from(KeyCode::Char('j')));
         assert_eq!(app.scroll, 1);
         assert_eq!(app.language, 0);
